+++ totp/types/plugin_state.h
@@ -19,7 +19,7 @@ typedef struct {
     float timezone_offset;
     ListNode* tokens_list;
     bool token_list_loaded;
-    uint8_t tokens_count;
+    uint16_t tokens_count;
 
     uint8_t* crypto_verify_data;
     size_t crypto_verify_data_length;
